chore: separates SkillProgress into it's own component
