+++ components/TheSkills.vue
@@ -1,47 +1,21 @@
 <script setup lang='ts'>
 import type { Skill } from '~/types'
 import resources from '~/data/resources.json'
-import levels from '~/data/levels.json'
 import { useGatherStore } from '~/composable/useGather'
-import { useSkillStore } from '~/composable/useSkills'
+import SkillProgress from '~/components/SkillProgress.vue'
 
 const gatherStore = useGatherStore()
-const skillStore = useSkillStore()
 const { progress, activeResource } = storeToRefs(gatherStore)
 const props = defineProps<{ skill: Skill }>()
 const filteredResources = computed(() => {
   return resources.filter(resource => resource.skillId === (props.skill as Skill).id)
 },
 )
-
-const remainingExp = computed(() => {
-  const nextLevel = levels.find(level => level.level === skillStore.activeSkill!.level + 1)
-  if (!nextLevel) return ''
-  return nextLevel.requiredXP - skillStore.activeSkill!.xp
-})
-const xpPerHour = computed(() => {
-  const gatheringTime = activeResource.value?.timeToGather || 0
-  const xpPerResource = activeResource.value?.experienceGiven || 0
-  const resourcesPerHour = 3600 / gatheringTime
-  const xpPerHour = xpPerResource * resourcesPerHour
-  if (!xpPerHour) {
-    return 0
-  }
-  else { return xpPerHour }
-})
-const nextLevelExpRequired = computed(() => {
-  const nextLevel = levels.find(level => level.level === skillStore.activeSkill!.level + 1)
-  if (!nextLevel) return ''
-  return nextLevel.requiredXP
-})
 </script>
 
 <template>
   <div>
-    Level: {{ skillStore.activeSkill!.level }}  <br>
-    Exp: {{ skillStore.activeSkill!.xp }}/ {{ nextLevelExpRequired }} <br>
-    Exp to next level: {{ remainingExp }} <br>
-    {{ xpPerHour }} xp/h
+    <SkillProgress :skill="skill" />
   </div>
   <progress
     :value="progress"
